서비스코드 기반 조회 추가

+++ src/main/java/com/example/tour/api/v1/response/TourInformationResponseDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
 @AllArgsConstructor
 public class TourInformationResponseDTO {
     @JsonProperty("region")
-    private Long code;
+    private Integer code;
 
     @JsonProperty("programs")
     private List<ProgramDTO> programs;
